services/backend/requirements.txt
# Core FastAPI and Server
fastapi==0.115.12
uvicorn==0.24.0
starlette==0.46.2
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4

# Authentication & Security
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Data Processing & Analysis
pandas==2.1.4
numpy==1.26.4
scipy==1.11.4
faker==20.1.0

# Machine Learning
scikit-learn==1.5.0
lightgbm==4.3.0
implicit==0.7.2
joblib==1.5.1
threadpoolctl==3.6.0
torch==2.1.2

# Association Rule Mining (for FBT recommendations)
mlxtend==0.23.1

# AWS Bedrock (for embeddings and LLM)
boto3==1.34.0
botocore==1.34.0

# Google Gemini (for image generation)
google-generativeai==0.8.3

# Anthropic (if needed for direct Claude API calls)
anthropic==0.57.1

# HTTP Clients
httpx==0.28.1
httpcore==1.0.9
httptools==0.6.4
aiohttp==3.12.0
aiohappyeyeballs==2.6.1
requests==2.32.3
requests-toolbelt==1.0.0

# Async & Concurrency
anyio==4.9.0
async-timeout==4.0.3
greenlet==3.2.2

# Validation & Serialization
pydantic==1.10.22
pydantic_core==2.33.2
email_validator==2.2.0
annotated-types==0.7.0
dataclasses-json==0.5.14
marshmallow==3.26.1

# Utilities
click==8.2.1
colorama==0.4.6
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
tqdm==4.67.1
tenacity==8.5.0
psutil==6.1.1

# Templates & Rendering
Jinja2==3.1.2
Mako==1.3.10
MarkupSafe==3.0.2

# Data Structures
attrs==25.3.0
frozenlist==1.6.0
multidict==6.4.4

# JSON & YAML
orjson==3.10.18
PyYAML==6.0.2
jsonpatch==1.33
jsonpointer==3.0.0

# Parsing & Validation
pyparsing==3.2.3
packaging==23.2
mypy_extensions==1.1.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2

# File I/O
h11==0.16.0
sniffio==1.3.1
watchfiles==1.0.5
websockets==12.0

# Cryptography & Security
rsa==4.9.1
ecdsa==0.19.1
pyasn1==0.6.1
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.0.7

# Networking
dnspython==2.7.0
yarl==1.20.0
aiosignal==1.3.2

# AWS & Cloud
s3transfer==0.9.0

# Utilities
six==1.17.0
jiter==0.10.0
jmespath==1.0.1
distro==1.9.0

# Caching & Performance
propcache==0.3.1

# Data Visualization & Analysis
graphviz==0.20.3
pygraphviz==1.14
grandalf==0.8

# Expression Evaluation
numexpr==2.10.2

# OpenAPI Schema Support
openapi-schema-pydantic==1.2.4

# Machine Learning / AI
transformers==4.46.3
huggingface-hub==0.26.2
